Fixed str_random helper. Don't allow deletion of admin users. Temp. fix for some display colours. phpMyAdmin env variable works.

+++ app/Http/Controllers/AdminController.php
@@ -123,6 +123,9 @@ class AdminController extends Controller
      */
     public function deleteUser(User $user)
     {
+	if ($user->isAdministrator()) {
+		return redirect('/userMaint')->with("error","Cannot delete administrator");
+	}
         $name = $user->fullname;
         $user->delete();
         return redirect('/userMaint')->with("error","User \"$name\" has been deleted");
